+++ engines_sort.h
@@ -18,6 +18,6 @@
 * @param count - количество записей в базе данных
 * @return нет
 */
-void gnome_sort(ENGINES *database, const unsigned count);
+void gnome_sort(ENGINES *database, const size_t count);
 
 #endif
